Refactor fait, a testé pour fixer les derniers problèmes

+++ src/backend/src/models/members/members.controller.ts
@@ -5,50 +5,70 @@ import {
   Get,
   Param,
   Put,
+  Req,
+  UnauthorizedException,
   UseGuards,
 } from '@nestjs/common';
 import { MembersService } from './members.service';
 import { UpdateMemberDto } from './dto/update.members.dto';
 import { DeleteResult, UpdateResult } from 'typeorm';
 import { AuthGuard } from '@nestjs/passport';
-import {
-  CheckPolicies,
-  PoliciesGuard,
-} from '../../security/policy/policy.guard';
-import {
-  DeleteMembersPolicyhandler,
-  ReadMembersPolicyhandler,
-  UpdateMembersPolicyhandler,
-} from '../../security/policy/handler/members.policyhandler';
 import { MemberDto, ToMember, ToMemberDto } from './dto/members.dto';
+import {
+  Action,
+  CaslAbilityFactory,
+} from '../../security/casl/casl-ability.factory';
 
 /**
  * Member controller
  */
 @Controller('members')
 export class MembersController {
-  constructor(private memberService: MembersService) {}
+  constructor(
+    private memberService: MembersService,
+    private caslAbilityFactory: CaslAbilityFactory,
+  ) {}
 
   /******************* GET    ************************/
   @Get('email/:email')
-  @UseGuards(AuthGuard('jwt'), PoliciesGuard)
-  @CheckPolicies(new ReadMembersPolicyhandler())
-  async findOneByEmail(@Param('email') email: string): Promise<MemberDto> {
-    return ToMemberDto(await this.memberService.findOne({ email: email }));
+  @UseGuards(AuthGuard('jwt'))
+  async findOneByEmail(
+    @Param('email') email: string,
+    @Req() req,
+  ): Promise<MemberDto> {
+    const ability = this.caslAbilityFactory.createForMember(req.user);
+    const member = await this.memberService.findOne({ email: email });
+
+    if (ability.can(Action.Read, member)) return ToMemberDto(member);
+
+    throw new UnauthorizedException();
   }
 
   /******************* PUT    ************************/
   @Put()
-  @UseGuards(AuthGuard('jwt'), PoliciesGuard)
-  @CheckPolicies(new UpdateMembersPolicyhandler())
-  async update(@Body() member: UpdateMemberDto): Promise<UpdateResult> {
-    return this.memberService.update(ToMember(member));
+  @UseGuards(AuthGuard('jwt'))
+  async update(
+    @Body() updatedMember: UpdateMemberDto,
+    @Req() req,
+  ): Promise<UpdateResult> {
+    const ability = this.caslAbilityFactory.createForMember(req.user);
+    const member = await this.memberService.findOne({ id: updatedMember.id });
+
+    if (ability.can(Action.Update, member)) {
+      return this.memberService.update(ToMember(member));
+    }
+    throw new UnauthorizedException();
   }
   /******************* DELETE ************************/
   @Delete(':uuid')
-  @UseGuards(AuthGuard('jwt'), PoliciesGuard)
-  @CheckPolicies(new DeleteMembersPolicyhandler())
-  async delete(@Param('uuid') uuid: string): Promise<DeleteResult> {
-    return this.memberService.delete(uuid);
+  @UseGuards(AuthGuard('jwt'))
+  async delete(@Param('uuid') uuid: string, @Req() req): Promise<DeleteResult> {
+    const ability = this.caslAbilityFactory.createForMember(req.user);
+    const member = await this.memberService.findOne({ id: uuid });
+
+    if (ability.can(Action.Delete, member)) {
+      return this.memberService.delete(uuid);
+    }
+    throw new UnauthorizedException();
   }
 }
