+++ src/backend/src/models/species/species.controller.ts
@@ -1,42 +1,40 @@
 import {
   Body,
-  ClassSerializerInterceptor,
   Controller,
   Delete,
   Get,
   Param,
   Post,
   Put,
+  Req,
+  UnauthorizedException,
   UseGuards,
-  UseInterceptors,
 } from '@nestjs/common';
 import { SpeciesService } from './species.service';
 import { AuthGuard } from '@nestjs/passport';
-import {
-  CheckPolicies,
-  PoliciesGuard,
-} from '../../security/policy/policy.guard';
 import { CreateSpeciesDto } from './dto/create.species.dto';
 import { Species } from './entities/species.entity';
 import { UpdateSpeciesDto } from './dto/update.species.dto.js';
 import { DeleteResult, UpdateResult } from 'typeorm';
-import {
-  CreateSpeciesPolicyhandler,
-  DeleteSpeciesPolicyhandler,
-  UpdateSpeciesPolicyhandler,
-} from '../../security/policy/handler/species.policyhandler';
 import { SpeciesDto, ToSpecies, ToSpeciesDto } from './dto/species.dto';
 import {
   ToTranslatedSpeciesDto,
   TranslatedSpeciesDto,
 } from './dto/translated.species.dto.js';
+import {
+  Action,
+  CaslAbilityFactory,
+} from '../../security/casl/casl-ability.factory';
 
 /**
  * Race controller
  */
 @Controller('species')
 export class SpeciesController {
-  constructor(private speciesService: SpeciesService) {}
+  constructor(
+    private speciesService: SpeciesService,
+    private caslAbilityFactory: CaslAbilityFactory,
+  ) {}
 
   /******************* GET    ************************/
 
@@ -47,8 +45,6 @@ export class SpeciesController {
     );
   }
 
-  // For example http://localhost:3000/species/fr
-  // Answer [{ "id": 25, "name": "chien" }, { "id": 25, "name": "chat" }, ... ]
   @Get(':lang')
   async findAllTranslated(
     @Param('lang') lang: string,
@@ -78,29 +74,43 @@ export class SpeciesController {
 
   /******************* POST   ************************/
   @Post()
-  @UseGuards(AuthGuard('jwt'), PoliciesGuard)
-  @CheckPolicies(new CreateSpeciesPolicyhandler())
-  async create(@Body() species: CreateSpeciesDto): Promise<SpeciesDto> {
-    return ToSpeciesDto(
-      await this.speciesService.createSpecies(
-        ToSpecies({ ...species, id: undefined }),
-      ),
-    );
+  @UseGuards(AuthGuard('jwt'))
+  async create(
+    @Body() species: CreateSpeciesDto,
+    @Req() req,
+  ): Promise<SpeciesDto> {
+    const ability = this.caslAbilityFactory.createForMember(req.user);
+
+    if (ability.can(Action.Create, Species)) {
+      return ToSpeciesDto(
+        await this.speciesService.createSpecies(
+          ToSpecies({ ...species, id: undefined }),
+        ),
+      );
+    }
+    throw new UnauthorizedException();
   }
 
   /******************* PUT    ************************/
   @Put()
-  @UseGuards(AuthGuard('jwt'), PoliciesGuard)
-  @CheckPolicies(new UpdateSpeciesPolicyhandler())
-  update(@Body() species: UpdateSpeciesDto): Promise<UpdateResult> {
-    return this.speciesService.updateSpecies(ToSpecies(species));
+  @UseGuards(AuthGuard('jwt'))
+  update(@Body() species: UpdateSpeciesDto, @Req() req): Promise<UpdateResult> {
+    const ability = this.caslAbilityFactory.createForMember(req.user);
+
+    if (ability.can(Action.Update, Species)) {
+      return this.speciesService.updateSpecies(ToSpecies(species));
+    }
+    throw new UnauthorizedException();
   }
 
   /******************* DELETE ************************/
   @Delete(':id')
-  @UseGuards(AuthGuard('jwt'), PoliciesGuard)
-  @CheckPolicies(new DeleteSpeciesPolicyhandler())
-  deleteOne(@Param('id') id: string): Promise<DeleteResult> {
-    return this.speciesService.deleteSpecies(parseInt(id));
+  @UseGuards(AuthGuard('jwt'))
+  deleteOne(@Param('id') id: string, @Req() req): Promise<DeleteResult> {
+    const ability = this.caslAbilityFactory.createForMember(req.user);
+    if (ability.can(Action.Delete, Species)) {
+      return this.speciesService.deleteSpecies(parseInt(id));
+    }
+    throw new UnauthorizedException();
   }
 }
